login

+++ app/Http/Controllers/TasksController.php
@@ -16,11 +16,27 @@ use App\Task;
          // getでtasks/にアクセスされた場合の「一覧表示処理」
         public function index()
         {
-            $tasks = Task::all();
-            return view('tasks.index',[
+            $data = [];
+            if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
                 'tasks' => $tasks,
-                ]);
+            ];
+            
         }
+
+            // Welcomeビューでそれらを表示
+            return view('welcome', $data);
+        }
+        
+
+     
     
         /**
          * Show the form for creating a new resource.
@@ -47,22 +63,23 @@ use App\Task;
          // postでtasks/にアクセスされた場合の「新規登録処理」
         public function store(Request $request)
         {
-            
             // バリデーション
             $request->validate([
                 'status' => 'required|max:10', 
                 'content' => 'required|max:255',
             ]);
-           // メッセージを作成
+            
+            // メッセージを作成
             $task = new Task;
             $task->status = $request->status;
             $task->content = $request->content;
+            $task->user_id = \Auth::id();
+            
             $task->save();
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 前のURLへリダイレクトさせる
+            return redirect('/');
         }
-    
         /**
          * Display the specified resource.
          *
@@ -75,12 +92,21 @@ use App\Task;
           // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+
+        // 自分のものかチェック
+        if (\Auth::id() === $task->user_id) {
+            // 自分のもの
+            return view('tasks.show', [
+            'task' => $task,
+        ]);
+        } else {
+            // 他人のもの
+           return back();
+        }
         return view('tasks.show', [
             'task' => $task,
         ]);
         }
-    
         /**
          * Show the form for editing the specified resource.
          *
@@ -94,11 +120,17 @@ use App\Task;
             // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            
+        if (\Auth::id() === $task->user_id) {
+            // 自分のもの
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        } else {
+            // 他人のもの
+            return back();
+        }
+        // メッセージ編集ビューでそれを表示
+        
         }
     
         /**
@@ -122,6 +154,7 @@ use App\Task;
         // メッセージを更新
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
+        
         $task->save();
 
         // トップページへリダイレクトさせる
@@ -134,15 +167,16 @@ use App\Task;
          * @param  int  $id
          * @return \Illuminate\Http\Response
          */
-          // deleteでmtasks/（任意のid）にアクセスされた場合の「削除処理」
+          // deleteでtasks/（任意のid）にアクセスされた場合の「削除処理」
         public function destroy($id)
         {
              // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+         if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        // トップページへリダイレクトさせる
+        // 前のURLへリダイレクトさせる
         return redirect('/');
         }
-    }
+}
